+++ src/init.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-function init::init() {
+function init::project() {
   local tests_dir="${1:-$BASHUNIT_DEFAULT_PATH}"
   mkdir -p "$tests_dir"