+++ src/env.sh
@@ -28,6 +28,7 @@ _DEFAULT_STOP_ON_FAILURE="false"
 _DEFAULT_SHOW_EXECUTION_TIME="true"
 _DEFAULT_VERBOSE="false"
 _DEFAULT_BENCH_MODE="false"
+_DEFAULT_NO_OUTPUT="false"
 
 : "${BASHUNIT_PARALLEL_RUN:=${PARALLEL_RUN:=$_DEFAULT_PARALLEL_RUN}}"
 : "${BASHUNIT_SHOW_HEADER:=${SHOW_HEADER:=$_DEFAULT_SHOW_HEADER}}"
@@ -37,6 +38,7 @@ _DEFAULT_BENCH_MODE="false"
 : "${BASHUNIT_SHOW_EXECUTION_TIME:=${SHOW_EXECUTION_TIME:=$_DEFAULT_SHOW_EXECUTION_TIME}}"
 : "${BASHUNIT_VERBOSE:=${VERBOSE:=$_DEFAULT_VERBOSE}}"
 : "${BASHUNIT_BENCH_MODE:=${BENCH_MODE:=$_DEFAULT_BENCH_MODE}}"
+: "${BASHUNIT_NO_OUTPUT:=${NO_OUTPUT:=$_DEFAULT_NO_OUTPUT}}"
 
 function env::is_parallel_run_enabled() {
   [[ "$BASHUNIT_PARALLEL_RUN" == "true" ]]
@@ -72,6 +74,10 @@ function env::is_verbose_enabled() {
 
 function env::is_bench_mode_enabled() {
   [[ "$BASHUNIT_BENCH_MODE" == "true" ]]
+}
+
+function env::is_no_output_enabled() {
+  [[ "$BASHUNIT_NO_OUTPUT" == "true" ]]
 }
 
 function env::active_internet_connection() {