+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix process time always shows as 0 ms
 - Fixed terminal width detection first tput and fall back stty
+- Refactor clock optimizing the implementation used to get the time
 
 ## [0.21.0](https://github.com/TypedDevs/bashunit/compare/0.20.0...0.21.0) - 2025-06-18